+++ n64/os/src/memory/memcpy.c
@@ -7,11 +7,11 @@
 
 /**
  * @file src/memory/memcpy.c
- * @brief Memory copying routine
+ * @brief Memory copying routines
  * @date 1 Aug 2018
  * @author pseudophpt
  *
- * This file provides routines a routine to copy memory from source to destination.
+ * This file provides routines to copy memory from source to destination.
  */
 
 #include <os_priv.h>
@@ -25,7 +25,6 @@
  * @author pseudophpt
  *
  * This function copies the region whose size is defined in the size parameter from the region at src to dest.
- *
  */
 void
 osCopyMemory
@@ -56,4 +55,42 @@ osCopyMemory
             continue;
         }
     }    
+}
+
+/**
+ * @brief Zero a memory region
+ * @param[out] dest Pointer to start of the region
+ * @param[in] size Size in bytes of region to zero
+ * @date 3 Aug 2018
+ * @author pseudophpt
+ *
+ * This function zeroes a memory region, typically one that will be used as BSS.
+ */
+void
+osZeroMemory
+(void *dest, int size) {
+    while (size > 0) {
+        // Must do byte write
+        if ((((u32)dest % 2) == 1) || ((size % 2) == 1)) {
+            *(u8 *)dest = 0;
+            dest = (void *)((u8 *)dest + 1);
+            size --;
+            continue;
+        }
+        // Must do hword write
+        else if ((((u32)dest % 4) == 2) || ((size % 4) == 2)) {
+            *(u16 *)dest = 0;
+            dest = (void *)((u16 *)dest + 1);
+            size -= 2;
+            continue;
+        }
+        // We can do a word write
+        else {
+            *(u32 *)dest = 0;
+            dest = (void *)((u32 *)dest + 1);
+            size -= 4;
+            continue;
+        }
+    }    
+
 }